requirements.txt
os
numpy 
pandas 
scipy.io 
python_speech_features 
keras.models
streamlit 
tensorflow 
tensorflow_hub 
librosa